--- src/main/java/com/hhplus/concert/api/balance/domain/Balance.java
@@ -1,29 +1,0 @@
-package com.hhplus.concert.api.balance.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-@Table(name = "BALANCE")
-public class Balance {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "BALANCE_ID")
-    private Long balanceId;
-    private Long userId;
-    private Long amount;
-    private LocalDateTime modDt;
-}
