--- src/main/java/com/hhplus/concert/api/balance/domain/User.java
@@ -1,29 +1,0 @@
-package com.hhplus.concert.api.balance.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-@Table(name = "USER")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "USER_ID")
-    private Long userId;
-    @Column(name = "UUID")
-    private String uuid;
-    @Column(name = "USER_NAME")
-    private String userName;
-}
