+++ src/main/java/com/hhplus/concert/api/concert/domain/repository/ConcertRepository.java
@@ -1,0 +1,8 @@
+package com.hhplus.concert.api.concert.domain.repository;
+
+import com.hhplus.concert.api.concert.domain.entity.Concert;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ConcertRepository extends JpaRepository<Concert, Long> {
+
+}
