--- src/main/java/com/hhplus/concert/api/concert/domain/repository/JpaScheduleRepository.java
@@ -1,10 +1,0 @@
-package com.hhplus.concert.api.concert.domain.repository;
-
-import com.hhplus.concert.api.concert.domain.entity.Schedule;
-import java.time.LocalDateTime;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaScheduleRepository extends JpaRepository<Schedule, Long> {
-    List<Schedule> findByConcertIdAndScheduleDateAfter(Long concertId, LocalDateTime now);
-}
