+++ src/main/java/com/hhplus/concert/api/concert/domain/repository/SeatRepository.java
@@ -1,0 +1,10 @@
+package com.hhplus.concert.api.concert.domain.repository;
+
+import com.hhplus.concert.api.concert.domain.entity.Seat;
+import com.hhplus.concert.api.concert.domain.type.SeatStatus;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SeatRepository extends JpaRepository<Seat, Long> {
+    List<Seat> findByScheduleIdAndSeatStatus(Long scheduleId, SeatStatus seatStatus);
+}
