+++ src/main/java/com/hhplus/concert/api/reservation/application/ReservationService.java
@@ -1,13 +1,13 @@
 package com.hhplus.concert.api.reservation.application;
 
-import com.hhplus.concert.api.auth.application.AuthService;
+import com.hhplus.concert.api.token.application.TokenService;
 import com.hhplus.concert.api.concert.application.ConcertService;
 import com.hhplus.concert.api.reservation.domain.entity.PaymentHistory;
 import com.hhplus.concert.api.reservation.domain.entity.Reservation;
 import com.hhplus.concert.api.reservation.domain.entity.ReservationSeat;
-import com.hhplus.concert.api.reservation.domain.repository.JpaPaymentHistoryRepository;
-import com.hhplus.concert.api.reservation.domain.repository.JpaReservationRepository;
-import com.hhplus.concert.api.reservation.domain.repository.JpaReservationSeatRepository;
+import com.hhplus.concert.api.reservation.domain.repository.PaymentHistoryRepository;
+import com.hhplus.concert.api.reservation.domain.repository.ReservationRepository;
+import com.hhplus.concert.api.reservation.domain.repository.ReservationSeatRepository;
 import com.hhplus.concert.api.reservation.domain.type.PaymentStatus;
 import com.hhplus.concert.api.reservation.domain.type.ReservationStatus;
 import com.hhplus.concert.exception.list.CustomBadRequestException;
@@ -23,23 +23,23 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class ReservationService {
 
-    private final JpaReservationRepository jpaReservationRepository;
-    private final JpaReservationSeatRepository jpaReservationSeatRepository;
-    private final JpaPaymentHistoryRepository jpaPaymentHistoryRepository;
-    private final AuthService authService;
+    private final ReservationRepository reservationRepository;
+    private final ReservationSeatRepository reservationSeatRepository;
+    private final PaymentHistoryRepository paymentHistoryRepository;
+    private final TokenService tokenService;
     private final ConcertService concertService;
 
     @Transactional
     public Reservation postReservationSeat(String uuid, Reservation reservation) {
-        authService.tokenStatusCheck(uuid);
+        tokenService.tokenStatusCheck(uuid);
         reservation.setConcertTitle(concertService.getConcertTitle(reservation.getConcertId()));
         reservation.setReservationStatus(ReservationStatus.SUCCESS);
         reservation.setReservationExpiry(LocalDateTime.now().plusMinutes(10));
         List<Long> seatIds = reservation.getReservationSeats().stream().map(i -> i.getSeatId()).collect(
             Collectors.toList());
         reservation.setTotalPrice(concertService.getSeatTotalPrice(seatIds));
-        jpaReservationRepository.save(reservation);
-        List<ReservationSeat> reservationSeats = jpaReservationSeatRepository.findAllBySeatIdInAndScheduleIdAndConcertId(seatIds,
+        reservationRepository.save(reservation);
+        List<ReservationSeat> reservationSeats = reservationSeatRepository.findAllBySeatIdInAndScheduleIdAndConcertId(seatIds,
                                                                                                       reservation.getScheduleId(),
                                                                                                       reservation.getConcertId());
         if (reservationSeats.size() != 0) {
@@ -55,7 +55,7 @@ public class ReservationService {
                                     .build());
             reservationSeats.add(seat);
         }
-        jpaReservationSeatRepository.saveAll(reservationSeats);
+        reservationSeatRepository.saveAll(reservationSeats);
 
         concertService.seatStatusImpossible(seatIds);
         return reservation;
@@ -63,16 +63,16 @@ public class ReservationService {
 
     @Transactional
     public PaymentHistory postPayment(String uuid, PaymentHistory paymentHistory) {
-        authService.tokenStatusCheck(uuid);
-        jpaReservationRepository.findByReservationIdAndUserId(
+        tokenService.tokenStatusCheck(uuid);
+        reservationRepository.findByReservationIdAndUserId(
             paymentHistory.getReservationId(), paymentHistory.getUserId()).orElseThrow(() -> {
             throw new CustomBadRequestException(HttpStatus.BAD_REQUEST, "예약 정보가 없습니다.");
         });
 
         paymentHistory.setPaymentStatus(PaymentStatus.SUCCESS);
         paymentHistory.setPaymentTime(LocalDateTime.now());
-        jpaPaymentHistoryRepository.save(paymentHistory);
-        authService.tokenExpired(uuid);
+        paymentHistoryRepository.save(paymentHistory);
+        tokenService.tokenExpired(uuid);
 
         return paymentHistory;
     }
