--- src/main/java/com/hhplus/concert/api/reservation/domain/repository/JpaPaymentHistoryRepository.java
@@ -1,8 +1,0 @@
-package com.hhplus.concert.api.reservation.domain.repository;
-
-import com.hhplus.concert.api.reservation.domain.entity.PaymentHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaPaymentHistoryRepository extends JpaRepository<PaymentHistory, Long> {
-
-}
