+++ src/main/java/com/hhplus/concert/api/reservation/domain/repository/PaymentHistoryRepository.java
@@ -1,0 +1,8 @@
+package com.hhplus.concert.api.reservation.domain.repository;
+
+import com.hhplus.concert.api.reservation.domain.entity.PaymentHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PaymentHistoryRepository extends JpaRepository<PaymentHistory, Long> {
+
+}
