+++ src/main/java/com/hhplus/concert/api/reservation/domain/repository/ReservationRepository.java
@@ -1,0 +1,9 @@
+package com.hhplus.concert.api.reservation.domain.repository;
+
+import com.hhplus.concert.api.reservation.domain.entity.Reservation;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+    Optional<Reservation> findByReservationIdAndUserId(Long reservationId, Long userId);
+}
