+++ src/main/java/com/hhplus/concert/api/token/presentation/dto/response/TokenGetResDTO.java
@@ -1,0 +1,23 @@
+package com.hhplus.concert.api.token.presentation.dto.response;
+
+import com.hhplus.concert.api.token.domain.type.TokenStatus;
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Schema(description = "대기열 토큰 조회 DTO")
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Builder
+public class TokenGetResDTO {
+    @Schema(description = "대기 번호")
+    private Long position;
+    @Schema(description = "토큰 상태")
+    private TokenStatus tokenStatus;
+    @Schema(description = "만료 시간")
+    private LocalDateTime expirationTime;
+}
